[Feat] SavePhotoScreen에서 Gemini 호출 후 LabelSerachScreen에서 그 상태를 관측하여 처리하도록 수정
