+++ app/src/main/java/com/and04/naturealbum/ui/savephoto/SavePhotoScreenPortrait.kt
@@ -39,7 +39,7 @@ fun SavePhotoScreenPortrait(
     isRepresented: State<Boolean>,
     onRepresentedChange: () -> Unit,
     photoSaveState: State<UiState<Unit>>,
-    onLabelSelect: (String) -> Unit,
+    onLabelSelect: () -> Unit,
     onBack: () -> Unit,
     savePhoto: (String, String, Label, Location, String, Boolean) -> Unit
 ) {
@@ -61,7 +61,6 @@ fun SavePhotoScreenPortrait(
 
         LabelSelection(
             label = label,
-            model = model,
             onClick = onLabelSelect,
         )
